Updated README.md for correct calling pattern
The `-c` and `-s` switches don't seem to be honored in the code itself, yet the README reflects this as if it actually works. Not using those switches allows the module to run. *Removed* those bogus switches (line 75).

Also - fixed some of the unneeded triple back-ticks (\`\`\`) inline. Markdown expects single back-ticks inline; triple back-ticks are for code blocks only.